+++ src/test/kotlin/ecommerce/controller/ProductControllerMockTest.kt
@@ -1,6 +1,6 @@
 package ecommerce.controller
 
-import ecommerce.TestFixture
+import ecommerce.BasicTestFixture
 import ecommerce.annotation.LoginMemberArgumentResolver
 import ecommerce.config.WebMvcConfiguration
 import ecommerce.dto.RegisteredMember
@@ -78,7 +78,7 @@ class ProductControllerMockTest
         fun `test getPages`() {
             setMockMemberTo(Role.USER)
 
-            val testProduct = TestFixture.createPaintingSadHuman()
+            val testProduct = BasicTestFixture.createPaintingSadHuman()
             whenever(productService.getPages(any(), any()))
                 .thenReturn(PageImpl(listOf(testProduct), PageRequest.of(0, 10), 20))
