+++ src/test/kotlin/ecommerce/service/CartServiceTest.kt
@@ -1,12 +1,12 @@
 package ecommerce.service
 
-import ecommerce.TestFixture.BRUSH
-import ecommerce.TestFixture.PAINTING_SAD_HUMAN
-import ecommerce.TestFixture.createAdmin
-import ecommerce.TestFixture.createMina
-import ecommerce.TestFixture.createPaintingHappyHuman
-import ecommerce.TestFixture.createPaintingSadHuman
-import ecommerce.TestFixture.createPetra
+import ecommerce.BasicTestFixture.BRUSH
+import ecommerce.BasicTestFixture.PAINTING_SAD_HUMAN
+import ecommerce.BasicTestFixture.createAdmin
+import ecommerce.BasicTestFixture.createMina
+import ecommerce.BasicTestFixture.createPaintingHappyHuman
+import ecommerce.BasicTestFixture.createPaintingSadHuman
+import ecommerce.BasicTestFixture.createPetra
 import ecommerce.dto.CartItemRequest
 import ecommerce.repository.MemberRepository
 import ecommerce.repository.ProductRepository
